+++ polaris-pipeline/DdeiClient/Domain/Args/Core/DdeiCaseIdentifiersArgDto.cs
@@ -2,6 +2,6 @@ namespace Ddei.Domain.CaseData.Args.Core
 {
     public class DdeiCaseIdentifiersArgDto : DdeiUrnArgDto
     {
-        public long CaseId { get; set; }
+        public int CaseId { get; set; }
     }
 }
