+++ polaris-pipeline/coordinator/TelemetryEvents/IndexedDocumentEvent.cs
@@ -13,7 +13,7 @@ namespace text_extractor.coordinator
 
         public Guid CorrelationId { get; set; }
         public string CaseUrn { get; set; }
-        public long CaseId { get; set; }
+        public int CaseId { get; set; }
         public string DocumentId { get; set; }
         public int? DocumentTypeId { get; set; }
         public string DocumentType { get; set; }
